style: rename middleware file
